refactor: separate `anims/timeline.py` into files in `timeline/`
