+++ janim/locale/zh_CN/LC_MESSAGES/janim.timeline.timeline.mo
