+++ janim/timeline/__init__.py
@@ -1,0 +1,8 @@
+# ruff: noqa
+from janim.timeline.timeline import (
+    Timeline,
+    BuiltTimeline,
+    TimelineItem,
+    PlaybackControl,
+    TimelinePlaybackControlItem,
+)
