+++ test/anims/test_composition.py
@@ -3,7 +3,7 @@ import unittest
 
 from janim.anims_core.animation import Animation
 from janim.anims.composition import Aligned, AnimGroup, Succession
-from janim.anims.timeline import Timeline
+from janim.timeline import Timeline
 from janim.anims_core.time import TimeRange
