+++ test/anims/test_transform.py
@@ -3,7 +3,7 @@ import unittest
 import numpy as np
 
 from janim.anims_core.animation import Animation
-from janim.anims.timeline import Timeline
+from janim.timeline import Timeline
 from janim.anims.transform import Transform
 from janim.anims_core.time import TimeRange
 from janim.constants import DOWN, LEFT, RIGHT, UP
